+++ R/dplyr_custom_functions.R
@@ -1,0 +1,40 @@
+
+#' Tibble Preview
+#' @description
+#' \lifecycle{experimental}
+#' Show a sample of all tibble data without hiding columns.
+#' @importFrom magrittr "%>%"
+#' @return A preview of a tibble.
+#' @export
+tp <- function(data, rows = 10) {
+  data <- dplyr::sample_n(data, size = rows)
+  print(data, n = Inf, width = Inf)
+}
+
+#' Ordered Factor case_when()
+#' @description
+#' \lifecycle{experimental}
+#' Can replace `case_when()` syntax and outputs an ordered factor in the same order as the cases, useful for meaningful ordering in plots and tables.  This is because for `case_when()` the arguments are evaluated in order, so you must proceed from the most specific to the most general. Tables and plots will therefor be ordered by the evaluation order.
+#' @param ... A sequence of two-sided formulas. See ?dplyr::case_when for details
+#' @return An ordered factor vector of length 1 or n, matching the length of the logical input or output vectors, with the type (and attributes) of the first RHS. Inconsistent lengths or types will generate an error.
+#' @importFrom magrittr "%>%"
+#' @export
+fct_case_when <- function(...) {
+  args <- as.list(match.call())
+  levels <- sapply(args[-1], function(f) f[[3]])  # extract RHS of formula
+  levels <- levels[!is.na(levels)]
+  ordered(dplyr::case_when(...), levels=levels)
+}
+
+#' Remove variables from tibble
+#' @description
+#' \lifecycle{stable}
+#' This is a simple negation of `dplyr::select`.
+#' @param .data A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.
+#' @param ... \<tidy-select\> One or more unquoted expressions separated by commas. Variable names can be used as if they were positions in the data frame, so expressions like x:y can be used to select a range of variables.
+#' @return An object of the same type as .data, with the specified columns removed.
+#' @importFrom magrittr "%>%"
+#' @export
+deselect <- function(.data, ...) {
+  dplyr::select(.data, -c(...))
+}